python/requirements.txt
faker
kafka-python
bson